feat: 完整测试环境工具和文档标准化
🎯 核心改进:
- 创建完整测试环境验证和设置工具链
- 重组测试目录结构(unit/integration/e2e)
- 建立强制性测试环境检查流程
- 同步更新所有相关README文档

🔧 新增测试工具:
- scripts/check_test_env.ps1: 30秒快速环境检查
- scripts/setup_test_env.ps1: 标准测试执行流程
- scripts/validate_test_config.py: 7步详细配置验证

📚 文档标准化:
- docs/development/testing-setup.md: 详细测试环境配置指南
- 更新MASTER.md: 强制测试环境检查点
- 更新testing-standards.md: 添加工具使用要求
- 同步所有README文档: 符合document-standards.md规范

🗂️ 测试重组:
- 按类型重组测试文件: unit/integration/e2e
- 更新conftest.py: 支持多种测试环境
- 删除重复和过时测试文件
- 创建各目录README导航文档

✅ MASTER规范合规:
- 所有文档创建触发README同步更新
- 强制测试环境检查点集成
- 文档交叉引用网络完整
- 测试工具标准化流程确立

+++ app/modules/quality_control/models.py
@@ -4,7 +4,7 @@
 定义证书管理相关的数据模型
 """
 
-from sqlalchemy import Column, String, DateTime, Boolean, Text, BigInteger
+from sqlalchemy import Column, String, DateTime, Boolean, Text, Integer
 from sqlalchemy.sql import func
 
 from app.core.database import Base
@@ -15,8 +15,8 @@ class Certificate(Base, TimestampMixin):
     """证书模型 - 质量控制证书管理"""
     __tablename__ = 'certificates'
 
-    # 主键 - 严格遵循docs/standards/database-standards.md规定
-    id = Column(BigInteger, primary_key=True, autoincrement=True, index=True)
+    # 主键 - 严格遵循docs/standards/database-standards.md规定：INTEGER主键
+    id = Column(Integer, primary_key=True, autoincrement=True, index=True)
     
     # 证书信息
     serial = Column(String(100), unique=True, nullable=False, index=True)
